backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "evolution-todo-backend"
version = "0.2.0"
description = "FastAPI backend for Evolution of Todo - Phase II"
authors = ["AI Team <ai@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlmodel = "^0.0.14"
pydantic = "^2.5.0"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0"
pyjwt = "^2.8.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
slowapi = "^0.1.9"
aiokafka = "^0.10.0"
openai-agents = "^0.0.15"
openai = "^1.58.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
